requirements.txt
python-dotenv~=1.2.1
instagrapi~=2.2.1